add configuration "arrowIsTwoCharWide"
